backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic-settings==2.5.2
python-multipart==0.0.9
qrcode[pil]==7.4.2
requests==2.32.3
